+++ app/controllers/api/saved_opportunities_controller.rb
@@ -3,20 +3,25 @@ class Api::SavedOpportunitiesController < ApiController
   # include DeviseControllerPatch
   before_action :set_saved_opportunity, only: [:show, :update, :destroy]
 
+  after_action :verify_authorized, except: :index
+  after_action :verify_policy_scoped, only: :index
+
   def index
-    @finalized_opportunities = @user.saved_opportunities
+    @saved_opportunities = policy_scope(SavedOpportunity)
 
     render :index
   end
 
   # GET /opportunities/1
   def show
+    authorize @saved_opportunity
     render :show
   end
 
   # POST /opportunities
   def create
     @saved_opportunity = SavedOpportunity.new(saved_opportunity_params)
+    authorize @saved_opportunity
 
     if @saved_opportunity.save
       # render json: @saved_opportunity, status: :created, location: @saved_opportunity
@@ -28,6 +33,7 @@ class Api::SavedOpportunitiesController < ApiController
 
   # PATCH/PUT /opportunities/1
   def update
+    authorize @saved_opportunity
     if @saved_opportunity.update(saved_opportunity_params)
       # render json: @saved_opportunity
       render :show
@@ -38,6 +44,7 @@ class Api::SavedOpportunitiesController < ApiController
 
   # DELETE /opportunities/1
   def destroy
+    authorize @saved_opportunity
     if @saved_opportunity.destroy
       render json: ['Saved opportunity was destroyed'], status: :ok
     else
